warheads sync command now fetches live source data and transforms into local files
